requirements.txt
langchain
langchain-community
openai
pinecone
python-dotenv
streamlit
